requirements.txt
flask
python-dotenv
tqdm
numpy
faster-whisper
pyannote.audio
torch
torchaudio
soundfile
ffmpeg-python
